Break petlogs into separate files by date and add READMEs
For older logs that didn't consistently use headings by date, their content has been
separated out by forensically reviewing the commits to their original files and
splitting out changes according to the date they were added. (This was especially messy
for the mk1 sturling logs, which was originally written to reflect a consistent "patch",
with all obsolete changes squashed away.)

This commit also un-deprecates the "mark 1" logs for stukilla by moving the "mark 2"
logs to the new name "stuzzy", and returning the old entries for stukilla to the same
group as the other machines.